refactor: MovieInfoDataLoader 구현, APIDataLoader 네이밍 BoxOfficeDataLoader로 수정
