+++ BoxOffice/Controller/BoxOfficeDataLoader.swift
@@ -1,5 +1,5 @@
 //
-//  APIDataLoader.swift
+//  BoxOfficeDataLoader.swift
 //  BoxOffice
 //
 //  Created by kaki, harry on 2023/04/05.
@@ -7,10 +7,10 @@
 
 import Foundation
 
-class APIDataLoader {
+class BoxOfficeDataLoader {
     private let networkManager = NetworkManager()
     
-    func loadDailyBoxOffice(completion: @escaping ((BoxOffice?, Error?)) -> ()) {
+    func loadDailyBoxOffice(completion: @escaping (BoxOffice?, Error?) -> ()) {
         let yesterdayText = DateFormatter.yesterdayText(format: .nonHyphen)
         let endPoint: BoxOfficeEndpoint = .fetchDailyBoxOffice(targetDate: yesterdayText)
         
@@ -19,9 +19,9 @@ class APIDataLoader {
             DispatchQueue.main.async {
                 switch result {
                 case .success(let data):
-                    completion((data, nil))
+                    completion(data, nil)
                 case .failure(let error):
-                    completion((nil, error))
+                    completion(nil, error)
                 }
             }
         }
